feat(subway) BaseAcceptanceTest 생성
