+++ src/test/java/nextstep/subway/acceptance/base/BaseAcceptanceTest.java
@@ -1,0 +1,21 @@
+package nextstep.subway.acceptance.base;
+
+import io.restassured.RestAssured;
+import org.junit.jupiter.api.BeforeEach;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.web.server.LocalServerPort;
+import org.springframework.test.context.jdbc.Sql;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+@Sql("/sql/truncate.sql")
+public abstract class BaseAcceptanceTest {
+    @LocalServerPort
+    int port;
+
+    @BeforeEach
+    void setUp() {
+        if (RestAssured.port == RestAssured.UNDEFINED_PORT) {
+            RestAssured.port = port;
+        }
+    }
+}
